+++ integration-test/sample-test/fixtures/api/clean_ns/a.clj
@@ -5,3 +5,5 @@
 a
 b
 c
+
+(ns api.clean-ns.a.other)