+++ integration-test/sample-test/src/api/clean_ns/a.clj
@@ -6,3 +6,5 @@
 a
 b
 c
+
+(ns api.clean-ns.a.other)